backend/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
psycopg2-binary
python-dotenv

